Save array's size instead of offset.

+++ be/src/olap/collection.h
@@ -25,16 +25,17 @@ namespace doris {
 struct Collection {
     // child column data
     void* data;
-    uint32_t length;
+    uint64_t length;
     // item has no null value if has_null is false.
     // item ```may``` has null value if has_null is true.
+    // null_count is better?
     bool has_null;
     // null bitmap
     bool* null_signs;
 
     Collection() : data(nullptr), length(0), has_null(false), null_signs(nullptr) {}
 
-    explicit Collection(uint32_t length)
+    explicit Collection(uint64_t length)
             : data(nullptr), length(length), has_null(false), null_signs(nullptr) {}
 
     Collection(void* data, size_t length)
@@ -46,7 +47,7 @@ struct Collection {
     Collection(void* data, size_t length, bool has_null, bool* null_signs)
             : data(data), length(length), has_null(has_null), null_signs(null_signs) {}
 
-    bool is_null_at(uint32_t index) { return this->has_null && this->null_signs[index]; }
+    bool is_null_at(uint64_t index) { return this->has_null && this->null_signs[index]; }
 
     bool operator==(const Collection& y) const;
     bool operator!=(const Collection& value) const;
